Refactor merchant relationship endpoint tests using one test file
